Adapted to new dataset. Fixed minor bugs.

+++ header/hypercube/AHypercube.h
@@ -21,7 +21,7 @@ public:
     virtual  vector<int> getBitVector(Item* item) = 0;
     vector<CubeEdge*> getHamingCLose(int hammingDist,CubeEdge* edge);
     void add(Item* item);
-    virtual double computeDistance(vector<int>&x,vector<int>&y) = 0;
+    virtual double computeDistance(vector<double>&x,vector<double>&y) = 0;
     pair<string,double> findCloser(Item *item,int max,int probes);
     vector<Item*> findRCloser(Item *item,int max,int probes,double radius);
     virtual int size() = 0;
