+++ src/lsh/cosineHashfunction.cpp
@@ -25,7 +25,7 @@ cosineHashfunction::~cosineHashfunction() {
 }
 
 int cosineHashfunction::hash(Item *item) {
-    vector<int>p_item = item->getContent();
+    vector<double>p_item = item->getContent();
     float sum=0;
     for(unsigned int i=0;i<p_item.size();i++){
         sum += p_item[i]*v[i];
